refactor: update app entry point and modernize architecture
- Move Event_TrackerApp to App directory
- Refactor to use StateObject for SessionManager and Router
- Add comprehensive documentation for AppDelegate and main app struct
- Remove direct auth check in favor of reactive session management
- Inject environment objects for proper state management
